refactor: clean up dependencies
