+++ packages/cspell-tools/package.json
@@ -45,14 +45,15 @@
   },
   "homepage": "https://github.com/streetsidesoftware/cspell#readme",
   "dependencies": {
+    "@cspell/cspell-pipe": "workspace:*",
     "commander": "^9.4.1",
     "cosmiconfig": "^7.0.1",
-    "cspell-io": "workspace:^",
-    "cspell-trie-lib": "workspace:^",
+    "cspell-io": "workspace:*",
+    "cspell-trie-lib": "workspace:*",
     "fs-extra": "^10.1.0",
     "gensequence": "^4.0.2",
     "glob": "^8.0.3",
-    "hunspell-reader": "workspace:^"
+    "hunspell-reader": "workspace:*"
   },
   "engines": {
     "node": ">=14"
